Inspect all `bldpkgs_dirs` that `conda-build` has
When the user does not provide an `inspection_directories` argument,
`conda-build-all` checks with `conda-build` for information about which
directory to look in by using the `bldpkgs_dir` config value. This
points to an OS specific build package directory. However this may
neglect other directories that `conda-build` might use like the `noarch`
directory.

In `conda-build` 1.19.1, the `bldpkgs_dirs` (with a final `s`) has been
around, which includes `bldpkgs_dir` and the `noarch` directory in a
`tuple`. This means we can comfortably use this with `conda-build` 1.x,
2.x, and 3.x without extra workarounds. So instead of setting
`inspection_directories` to `bldpkgs_dir` when undefined. Instead use
`bldpkgs_dirs` to get the platform specific an `noarch` directories.
This will avoid repeated rebuilds of `noarch` packages when they have
already been built once.

ref: https://github.com/conda/conda-build/commit/ab60a621694547ddc6c9563d1206b189f5bfc850